+++ R/article.R
@@ -9,8 +9,8 @@
 #' (e.g., `keep_tex = TRUE`).
 #'
 #' @param
-#' ...,keep_tex,latex_engine,citation_package,highlight,fig_caption,md_extensions,template,pandoc_args
-#' Arguments passed to `rmarkdown::[pdf_document]()`.
+#' ...,number_sections,keep_tex,latex_engine,citation_package,highlight,fig_caption,md_extensions,template,pandoc_args
+#' Arguments passed to [rmarkdown::pdf_document()].
 #' @section Details: You can find more details about each output format below.
 #' @name acm_article
 #' @rdname article
@@ -23,7 +23,8 @@ NULL
 #' @return An R Markdown output format.
 #' @examples \dontrun{
 #' rmarkdown::draft("MyArticle.Rmd", template = "acm", package = "rticles")
-#' rmarkdown::draft("MyArticle.Rmd", template = "asa", package = "rticles")}
+#' rmarkdown::draft("MyArticle.Rmd", template = "asa", package = "rticles")
+#' }
 #' @export
 #' @rdname article
 acm_article <- function(...) {
@@ -32,14 +33,15 @@ acm_article <- function(...) {
 
 #' @section `acs_article`: Format for creating an American Chemical Society
 #'   (ACS) Journal articles. Adapted from
-#'   <https://pubs.acs.org/page/4authors/submission/tex.html>.
+#'   `https://pubs.acs.org/page/4authors/submission/tex.html`
 #' @export
 #' @rdname article
-acs_article <- function(
-  ..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"), fig_caption = TRUE
-) {
+acs_article <- function(..., keep_tex = TRUE,
+                        md_extensions = c("-autolink_bare_uris"),
+                        fig_caption = TRUE) {
   pdf_document_format(
-    "acs", keep_tex = keep_tex, md_extensions = md_extensions,
+    "acs",
+    keep_tex = keep_tex, md_extensions = md_extensions,
     fig_caption = fig_caption, ...
   )
 }
@@ -48,23 +50,25 @@ acs_article <- function(
 #'   Economic Association (AER, AEJ, JEL, PP).
 #' @export
 #' @rdname article
-aea_article <- function(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris")) {
+aea_article <- function(..., keep_tex = TRUE,
+                        md_extensions = c("-autolink_bare_uris")) {
   pdf_document_format(
-    "aea", keep_tex = keep_tex, md_extensions = md_extensions, ...
+    "aea",
+    keep_tex = keep_tex, md_extensions = md_extensions, ...
   )
 }
 
 #' @section `agu_article`: Format for creating a American Geophysical Union
 #'   (AGU) article. Adapted from
-#'   <https://www.agu.org/Publish-with-AGU/Publish/#1>.
+#'   <https://www.agu.org/publish-with-agu/publish#1>.
 #' @export
 #' @rdname article
-agu_article <- function(
-  ..., keep_tex = TRUE, citation_package = 'natbib',
-  highlight = NULL, md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
-) {
+agu_article <- function(..., keep_tex = TRUE,
+                        citation_package = "natbib", highlight = NULL,
+                        md_extensions = c("-autolink_bare_uris", "-auto_identifiers")) {
   pdf_document_format(
-    "agu", keep_tex = keep_tex, highlight = highlight,
+    "agu",
+    keep_tex = keep_tex, highlight = highlight,
     citation_package = citation_package, md_extensions = md_extensions, ...
   )
 }
@@ -73,12 +77,11 @@ agu_article <- function(
 #'   l'AMQ.
 #' @export
 #' @rdname article
-amq_article <- function(
-  ..., latex_engine = 'xelatex', keep_tex = TRUE, fig_caption = TRUE,
-  md_extensions = c("-autolink_bare_uris")
-) {
+amq_article <- function(..., latex_engine = "xelatex", keep_tex = TRUE,
+                        fig_caption = TRUE, md_extensions = c("-autolink_bare_uris")) {
   pdf_document_format(
-    "amq", latex_engine = latex_engine, highlight = NULL, keep_tex = keep_tex,
+    "amq",
+    latex_engine = latex_engine, highlight = NULL, keep_tex = keep_tex,
     md_extensions = md_extensions, fig_caption = fig_caption, ...
   )
 }
@@ -95,13 +98,17 @@ ams_article <- function(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_
 }
 
 #' @section `asa_article`: This format was adapted from The American
-#'   Statistican (TAS) format, but it should be fairly consistent across
+#'   Statistician (TAS) format, but it should be fairly consistent across
 #'   American Statistical Association (ASA) journals.
 #' @export
 #' @rdname article
-asa_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
+asa_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
+  if (citation_package == "biblatex") {
+    stop("ASA template does not support `biblatex` for citation processing.")
+  }
   pdf_document_format(
-    "asa", keep_tex = keep_tex, citation_package = citation_package, ...
+    "asa",
+    keep_tex = keep_tex, citation_package = citation_package, ...
   )
 }
 
@@ -112,7 +119,8 @@ asa_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
 #' @rdname article
 arxiv_article <- function(..., keep_tex = TRUE) {
   pdf_document_format(
-    "arxiv", keep_tex = keep_tex, ...
+    "arxiv",
+    keep_tex = keep_tex, ...
   )
 }
 
@@ -120,10 +128,11 @@ arxiv_article <- function(..., keep_tex = TRUE) {
 #' <https://academic.oup.com/bioinformatics/pages/submission_online>.
 #' @export
 #' @rdname article
-bioinformatics_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
+bioinformatics_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
   pdf_document_format(
-    "bioinformatics", keep_tex = keep_tex, citation_package = citation_package,
-    md_extensions = "-auto_identifiers",...
+    "bioinformatics",
+    keep_tex = keep_tex, citation_package = citation_package,
+    md_extensions = "-auto_identifiers", ...
   )
 }
 
@@ -131,9 +140,10 @@ bioinformatics_article <- function(..., keep_tex = TRUE, citation_package = 'nat
 #'   Biometrics journal.
 #' @export
 #' @rdname article
-biometrics_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
+biometrics_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
   pdf_document_format(
-    'biometrics', keep_tex = keep_tex, citation_package = citation_package, ...
+    "biometrics",
+    keep_tex = keep_tex, citation_package = citation_package, ...
   )
 }
 
@@ -144,9 +154,10 @@ biometrics_article <- function(..., keep_tex = TRUE, citation_package = 'natbib'
 #'   function `ctex` is an alias of `ctex_article`.
 #' @export
 #' @rdname article
-ctex_article <- function(..., template = 'default', latex_engine = 'xelatex') {
+ctex_article <- function(..., template = "default", latex_engine = "xelatex") {
   pdf_document_format(
-    'ctex', latex_engine = latex_engine, template = template, ...
+    "ctex",
+    latex_engine = latex_engine, template = template, ...
   )
 }
 
@@ -156,14 +167,24 @@ ctex <- ctex_article
 
 #' @section `elsevier_article`: Format for creating submissions to Elsevier
 #'   journals. Adapted from
-#'   <https://www.elsevier.com/authors/policies-and-guidelines/latex-instructions>.
+#'   <https://www.elsevier.com/researcher/author/policies-and-guidelines/latex-instructions>.
+#'
+#' It requires a minimum version of 2.10 for Pandoc.
 #' @export
 #' @rdname article
-elsevier_article <- function(
-  ..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris")
-) {
+elsevier_article <- function(..., keep_tex = TRUE,
+                             md_extensions = c("-autolink_bare_uris"),
+                             citation_package = "natbib") {
+  if (citation_package == "biblatex") {
+    stop("Elsevier template does not support `biblatex` for citation processing.")
+  }
+  if (!rmarkdown::pandoc_available("2.10")) {
+    stop("`elsevier_article()` now requires a minimum of pandoc 2.10.")
+  }
   pdf_document_format(
-    "elsevier", keep_tex = keep_tex, md_extensions = md_extensions, ...
+    "elsevier",
+    keep_tex = keep_tex, md_extensions = md_extensions,
+    citation_package = citation_package, ...
   )
 }
 
@@ -172,8 +193,12 @@ elsevier_article <- function(
 #'   <https://www.frontiersin.org/about/author-guidelines>.
 #' @export
 #' @rdname article
-frontiers_article <- function(..., keep_tex = TRUE) {
-  pdf_document_format("frontiers", keep_tex = keep_tex, ...)
+frontiers_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
+  # check all arguments for format's default
+  if (citation_package != "natbib" && rmarkdown::pandoc_available("3.1.7")) {
+    stop("Frontiers template only supports 'natbib' for citation processing. Pandoc citeproc is no more compatible with the Frontier documentclass.")
+  }
+  pdf_document_format("frontiers", keep_tex = keep_tex, citation_package = citation_package,...)
 }
 
 
@@ -184,9 +209,20 @@ frontiers_article <- function(..., keep_tex = TRUE) {
 #' @export
 #' @rdname article
 glossa_article <- function(..., keep_tex = TRUE, latex_engine = "xelatex") {
-  pdf_document_format(
-    "glossa", keep_tex = keep_tex, latex_engine = latex_engine, ...
+  format <- pdf_document_format(
+    "glossa",
+    keep_tex = keep_tex, latex_engine = latex_engine, ...
   )
+  if (tinytex::is_tinytex() && tinytex::check_installed("microtype")) {
+    # TODO: known conflict - remove when fixed
+    tinytex::tlmgr_remove("microtype")
+    fun <- format$on_exit
+    format$on_exit <- function() {
+      if (is.function(fun)) fun()
+      if (!tinytex::check_installed("microtype")) tinytex::tlmgr_install("microtype")
+    }
+  }
+  format
 }
 
 #' @param journal one of `"aoas"`, `"aap"`, `"aop"`, `"aos"`, `"sts"` for `ims_article`
@@ -208,7 +244,6 @@ ims_article <- function(journal = c("aoas", "aap", "aop", "aos", "sts"),
                         md_extensions = c(
                           "-autolink_bare_uris" # disables automatic links
                         ), pandoc_args = NULL, ...) {
-
   journal <- match.arg(journal)
 
   with_kwsc <- journal %in% c("aap", "aop", "aos") # with keyword_subclass
@@ -216,17 +251,48 @@ ims_article <- function(journal = c("aoas", "aap", "aop", "aos", "sts"),
   args <- c(
     "journal" = journal,
     if (with_kwsc) c("with_kwsc" = with_kwsc)
-    )
+  )
 
   # Convert to pandoc arguments
-  pandoc_arg_list <- mapply(rmarkdown::pandoc_variable_arg, names(args), args,
-                            SIMPLIFY = FALSE, USE.NAMES = FALSE)
-  pandoc_arg_list <- unlist(pandoc_arg_list)
+  pandoc_arg_list <- list_to_pandoc_variable_args(args)
 
   pdf_document_format(
-    "ims", keep_tex = keep_tex, citation_package = citation_package,
+    "ims",
+    keep_tex = keep_tex, citation_package = citation_package,
     md_extensions = md_extensions, pandoc_args = c(pandoc_arg_list, pandoc_args),
     ...
+  )
+}
+
+#' @section `informs_article`: Format for creating submissions to
+#' INFORMS journals. Adapted from \samp{https://pubsonline.informs.org/authorportal/latex-style-files}.
+#'
+#' It requires a minimum version of 2.10 for Pandoc.
+#' @export
+#' @rdname article
+informs_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
+  if (citation_package != "natbib") {
+    stop("INFORMS template only supports `natbib` for citation processing.")
+  }
+
+  if (!rmarkdown::pandoc_available("2.10")) {
+    stop("informs_article requires a minimum of pandoc 2.10.")
+  }
+  pdf_document_format(
+    "informs",
+    keep_tex = keep_tex, citation_package = citation_package, ...
+  )
+}
+
+#' @section `iop_article`: Format for creating submissions to
+#'   IOP journals. Adapted from \samp{https://publishingsupport.iopscience.iop.org/questions/latex-template/}.
+#'   Please read the guidelines at this link when preparing your article.
+#' @export
+#' @rdname article
+iop_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
+  pdf_document_format(
+    "iop",
+    keep_tex = keep_tex, citation_package = citation_package, ...
   )
 }
 
@@ -235,11 +301,11 @@ ims_article <- function(journal = c("aoas", "aap", "aop", "aos", "sts"),
 #'   <https://acousticalsociety.org/preparing-latex-manuscripts/>.
 #' @export
 #' @rdname article
-jasa_article <- function(
-  ..., keep_tex = TRUE, latex_engine = "xelatex", citation_package = "natbib"
-) {
+jasa_article <- function(..., keep_tex = TRUE, latex_engine = "xelatex",
+                         citation_package = "natbib") {
   pdf_document_format(
-    "jasa", keep_tex = keep_tex, latex_engine = latex_engine,
+    "jasa",
+    keep_tex = keep_tex, latex_engine = latex_engine,
     citation_package = citation_package, ...
   )
 }
@@ -253,22 +319,37 @@ jasa_article <- function(
 #'   Project Public License (LPPL), Version 1.3c.
 #' @export
 #' @rdname article
-lipics_article <- function(
-  ..., latex_engine = 'xelatex', # xelatex used for 'thin space' Unicode
-                                 # character, see YAML field 'authorrunning'
-  keep_tex = TRUE, citation_package = "natbib", md_extensions = c(
-    "-autolink_bare_uris", # disables automatic links
-    "-auto_identifiers"    # disables \hypertarget commands
-  )
-) {
+lipics_article <- function(..., latex_engine = "xelatex", # xelatex used for 'thin space' Unicode
+                           # character, see YAML field 'authorrunning'
+                           keep_tex = TRUE, citation_package = "natbib",
+                           md_extensions = c(
+                             "-autolink_bare_uris", # disables automatic links
+                             "-auto_identifiers" # disables \hypertarget commands
+                           )) {
   # quick dev shortcut for Ubuntu: click "Install and restart" then run:
   # unlink("MyArticle/", recursive = TRUE); rmarkdown::draft("MyArticle.Rmd", template = "lipics", package = "rticles", edit = FALSE); rmarkdown::render("MyArticle/MyArticle.Rmd"); system(paste0("xdg-open ", here::here("MyArticle", "MyArticle.pdf")))
   pdf_document_format(
-    "lipics", latex_engine = latex_engine,
+    "lipics",
+    latex_engine = latex_engine,
     citation_package = citation_package, keep_tex = keep_tex,
     md_extensions = md_extensions, ...
   )
 }
+
+
+#' @section `lncs_article`: Format for creating submissions to
+#'   LNCS - Lecture Notes in Computer Science - articles.
+#'   Adapted from the official Instructions for Authors at
+#'   <https://www.springer.com/gp/computer-science/lncs/conference-proceedings-guidelines>
+#'   and the  template from the archive `LaTeX2e+Proceedings+Templates+download.zip` downloaded
+#'   with version tag 2.21.
+#' @export
+#' @rdname article
+lncs_article <- function(..., keep_tex = TRUE, citation_package = c("default", "natbib")) {
+  citation_package <- match.arg(citation_package)
+  pdf_document_format("lncs", keep_tex = keep_tex, citation_package = citation_package, ...)
+}
+
 
 #' @section `jedm_article`: Format for creating Journal of Educational
 #'   Data Mining (JEDM) articles. Adapted from
@@ -277,19 +358,9 @@ lipics_article <- function(
 #' @rdname article
 jedm_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
   pdf_document_format("jedm",
-                      keep_tex = keep_tex,
-                      citation_package = citation_package,
-                      ...)
-}
-
-#' @section `mdpi_article`: Format for creating submissions to
-#'   Multidisciplinary Digital Publishing Institute (MDPI) journals. Adapted
-#'   from <https://www.mdpi.com/authors/latex>.
-#' @export
-#' @rdname article
-mdpi_article <- function(..., keep_tex = TRUE) {
-  pdf_document_format(
-    "mdpi", keep_tex = keep_tex, citation_package = "natbib", ...
+    keep_tex = keep_tex,
+    citation_package = citation_package,
+    ...
   )
 }
 
@@ -300,23 +371,8 @@ mdpi_article <- function(..., keep_tex = TRUE) {
 #' @rdname article
 mnras_article <- function(..., keep_tex = TRUE, fig_caption = TRUE) {
   pdf_document_format(
-    "mnras", keep_tex = keep_tex, fig_caption = fig_caption, ...
-  )
-}
-
-#' @section `oup_article`: Format for creating submissions to many Oxford University Press
-#'   journals. Adapted from
-#'   <https://academic.oup.com/journals/pages/authors/preparing_your_manuscript>
-#'   and <https://academic.oup.com/icesjms/pages/General_Instructions>.
-#' @export
-#' @rdname article
-oup_article <- function(
-  ..., keep_tex = TRUE,
-  md_extensions = c("-autolink_bare_uris")
-) {
-  pdf_document_format(
-    "oup",
-    keep_tex = keep_tex, md_extensions = md_extensions, ...
+    "mnras",
+    keep_tex = keep_tex, fig_caption = fig_caption, ...
   )
 }
 
@@ -327,7 +383,7 @@ oup_article <- function(
 #' @export
 #' @rdname article
 peerj_article <- function(..., keep_tex = TRUE) {
-  pdf_document_format("peerj",  keep_tex = keep_tex, ...)
+  pdf_document_format("peerj", keep_tex = keep_tex, ...)
 }
 
 #' @section `pihph_article`: Format for creating submissions to the Papers
@@ -340,24 +396,24 @@ peerj_article <- function(..., keep_tex = TRUE) {
 #'   template.
 #' @export
 #' @rdname article
-pihph_article <- function(
-  ..., keep_tex = TRUE, latex_engine = "xelatex",
-  citation_package = "biblatex"
-) {
+pihph_article <- function(..., keep_tex = TRUE, latex_engine = "xelatex",
+                          citation_package = "biblatex") {
   pdf_document_format(
-    "pihph", keep_tex = keep_tex, latex_engine = latex_engine,
-    citation_package = citation_package, ...)
+    "pihph",
+    keep_tex = keep_tex, latex_engine = latex_engine,
+    citation_package = citation_package, ...
+  )
 }
 
 #' @section `plos_article`: Format for creating submissions to PLOS
 #'   journals. Adapted from <https://journals.plos.org/ploscompbiol/s/latex>.
 #' @export
 #' @rdname article
-plos_article <- function(
-  ..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris")
-) {
+plos_article <- function(..., keep_tex = TRUE,
+                         md_extensions = c("-autolink_bare_uris")) {
   pdf_document_format(
-    "plos", keep_tex = keep_tex, md_extensions = md_extensions, ...
+    "plos",
+    keep_tex = keep_tex, md_extensions = md_extensions, ...
   )
 }
 
@@ -370,8 +426,8 @@ pnas_article <- function(..., keep_tex = TRUE) {
 }
 
 #' @section `sage_article`: Format for creating submissions to Sage
-#'   Journals. Based on the official Sage Journals
-#'   \verb{https://uk.sagepub.com/sites/default/files/sage_latex_template_4.zip}{class}.
+#'   Journals. Based on the official Sage Journals Class.
+#'   Available at `https://uk.sagepub.com/sites/default/files/sage_latex_template_4.zip`.
 #'
 #' Possible arguments for the YAML header are:
 #' * `title` title of the manuscript
@@ -390,13 +446,16 @@ pnas_article <- function(..., keep_tex = TRUE) {
 #' @rdname article
 sage_article <- function(..., highlight = NULL, citation_package = "natbib") {
   pdf_document_format(
-    "sage", highlight = highlight, citation_package = citation_package, ...
+    "sage",
+    highlight = highlight, citation_package = citation_package, ...
   )
 }
 
 #' @section `sim_article`: Format for creating submissions to Statistics in
 #'   Medicine. Based on the official Statistics in Medicine
-#'   [class](http://onlinelibrary.wiley.com/journal/10.1002/(ISSN)1097-0258/homepage/la_tex_class_file.htm).
+#'   at `https://authorservices.wiley.com/author-resources/Journal-Authors/Prepare/new-journal-design.html`.
+#'
+#'   This format uses xelatex by default as PDF engine to support the specific NJD fonts, per guideline.
 #'
 #' Possible arguments for the YAML header are:
 #' * `title` title of the manuscript
@@ -408,16 +467,18 @@ sage_article <- function(..., highlight = NULL, citation_package = "natbib") {
 #' * `received`, `revised`, `accepted` dates of submission, revision, and acceptance of the manuscript
 #' * `abstract` abstract, limited to 250 words
 #' * `keywords` up to 6 keywords
+#' * `abbreviations`, list of abbreviations and description separated by a comma
 #' * `bibliography` BibTeX `.bib` file
-#' * `classoption` options of the `WileyNJD-v2` class
+#' * `classoption` options of the `WileyNJD` class
 #' * `longtable` set to `true` to include the `longtable` package, used by default from `pandoc` to convert markdown to LaTeX code
 #' * `header-includes`: custom additions to the header, before the `\begin{document}` statement
 #' * `include-after`: for including additional LaTeX code before the `\end{document}` statement
 #' @export
 #' @rdname article
-sim_article <- function(..., highlight = NULL, citation_package = "natbib") {
+sim_article <- function(..., highlight = NULL, citation_package = "natbib", latex_engine = "xelatex") {
   pdf_document_format(
-    "sim", highlight = highlight, citation_package = citation_package, ...
+    "sim",
+    highlight = highlight, citation_package = citation_package, latex_engine = latex_engine, ...
   )
 }
 
@@ -425,18 +486,142 @@ sim_article <- function(..., highlight = NULL, citation_package = "natbib") {
 #'   Macro package for Springer Journals.
 #' @export
 #' @rdname article
-springer_article <- function(..., keep_tex = TRUE, citation_package = 'default'){
-  pdf_document_format(
-    "springer", keep_tex = keep_tex, citation_package = citation_package, ...
+springer_article <- function(..., keep_tex = TRUE,  citation_package = "natbib",
+                             number_sections = TRUE, latex_engine = "pdflatex",
+                             pandoc_args = NULL) {
+
+  if (!rmarkdown::pandoc_available("2.11.4")) {
+    stop("`springer_article()` now requires a minimum of pandoc 2.11.4")
+  }
+
+  if (citation_package == "biblatex") {
+    stop("'springer_article' does not support `biblatex` for citation processing. Use 'natbib' instead.")
+  }
+
+  # to compile with pdflatex/xelatex
+  # use pdflatex option in the pandoc's template for the document class
+  if(latex_engine %in% c("pdflatex", "xelatex")) {
+    pandoc_args <- c(pandoc_args, rmarkdown::pandoc_variable_arg("pdflatex"))
+  }
+
+  if(!number_sections) {
+    pandoc_args <- c(pandoc_args, rmarkdown::pandoc_variable_arg("unnumbered"))
+  }
+
+  format <- pdf_document_format(
+    "springer",
+    keep_tex = keep_tex,
+    citation_package = citation_package,
+    number_sections = number_sections,
+    latex_engine = latex_engine,
+    pandoc_args = pandoc_args, ...
   )
+
+  pre_knit_fun <- format$pre_knit
+  format$pre_knit <-  function(input, ...) {
+    if (is.function(pre_knit_fun)) pre_knit_fun(input, ...)
+    # for backward compatibility as we changed the template in
+    # https://github.com/rstudio/rticles/pull/494
+    options <- rmarkdown::yaml_front_matter(input)
+    new_template_msg <- function(ver) return (c(sprintf("If you are rendering an old Rmd, be advise that the template has changed in version '%s'\n", ver),
+    " and you should check new template or start from a fresh template to get latest resources and new YAML header format."))
+    if (is.null(options[["classoptions"]])) {
+      stop("`springer_article()` now requires the 'classoptions' field in YAML front matter. ",
+           new_template_msg("0.25"), call. = FALSE)
+    } else if ("sn-mathphys" %in% options[["classoptions"]]) {
+      stop("classoptions `sn-mathphys` detected. `springer_article()` now uses 'sn-mathphys-num' or `sn-mathphys-ay`. ",
+              new_template_msg("0.28"), call. = FALSE)
+    }
+    if (!is.null(options[["biblio-style"]])) {
+      warning("`springer_article()` now ignores the 'biblio-style' field in YAML front matter. ",
+      "Reference style for 'natbib' is now set using the 'classoptions' field.\n",
+      new_template_msg("0.25"))
+    }
+    if (!is.null(options[["PACS"]])) {
+      warning("`springer_article()` now ignores the 'PACS' field in YAML front matter to use `pacs.jel` and `pacs.msc`. ",
+              new_template_msg("0.25"))
+    }
+    if (!is.null(options[["authors"]][["name"]])) {
+      stop("`springer_article()` now uses different authors and affiliations fields.\n",
+              new_template_msg("0.25"), call. = FALSE)
+    }
+    return(invisible(NULL))
+  }
+  format
 }
 
 #' @section `tf_article`: Format for creating submissions to a Taylor & Francis journal. Adapted from
 #' \samp{https://www.tandf.co.uk/journals/authors/InteractCADLaTeX.zip}.
 #' @export
 #' @rdname article
-tf_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
+tf_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
   pdf_document_format(
-    "tf", keep_tex = keep_tex, citation_package = citation_package, ...
+    "tf",
+    keep_tex = keep_tex, citation_package = citation_package, ...
+  )
+}
+
+#' @section \code{trb_article}: Format for creating submissions to the Transportation
+#'   Research Board Annual Meeting. Adapted from
+#'   \samp{https://www.overleaf.com/latex/templates/transportation-research-board-trb-latex-template/jkfndnnkkksw},
+#'   which in turn is hosted at \samp{https://github.com/chiehrosswang/TRB_LaTeX_tex}
+#' @export
+#' @rdname article
+trb_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
+  pdf_document_format(
+    "trb",
+    keep_tex = keep_tex, citation_package = citation_package, ...
+  )
+}
+
+#' @section `wellcomeor_article`: Format for creating submissions to
+#' Wellcome Open Research. Adapted from <overleaf.com/latex/templates/wellcome-open-research-article-template/hsmhhbpxvvbj>.
+#' @export
+#' @rdname article
+wellcomeor_article <- function(..., number_sections = FALSE, keep_tex = TRUE,
+                               citation_package = "natbib") {
+  pdf_document_format(
+    "wellcomeor",
+    keep_tex = keep_tex, number_sections = number_sections,
+    citation_package = citation_package, ...
+  )
+}
+
+#' @section `isba_article`: Format for creating submissions to Bayesian analysis.
+#' Based on the official Bayesian analysis [class](https://www.e-publications.org/isba/support/).
+#' Template shows how to use this format as a base format for `bookdown::pdf_book`, but it can very well be used on its own (with limitations that figure referencing will not work).
+#' Note that the template sets `md_extensions` to exclude `-autolink_bare_uris` because otherwise author emails produce error
+#'
+#' Possible arguments for the YAML header are:
+#' * `title` title of the manuscript. Shorter version of the title can be provided as `runtitle`.
+#' * `classoption` should equal `ba` or `ba,preprint` for supplementary article.``
+#' * `author` list of authors, containing `firstname`, `lastname`, `email`, `url`, `affiliationref` (as code) and `footnoterefs` (as list of codes)
+#' * `affiliations` list containing `ref` (code for defining `author` affiliations), institution `name` and `address` itself
+#' * `footnotes` a list of two-element entries: `ref` and `text`
+#' * `abstract` abstract, limited to 250 words
+#' * `MSC2020primary`, `MSC2020primary` lists of codes from [MCS2020 database](https://mathscinet.ams.org/mathscinet/msc/msc2020.html)
+#' * `keywords` a list of keywords
+#' * `supplements` a list of entries with two elements `title` and `description`
+#' * `doi` DOI of the article
+#' * `arxiv` Arxiv id
+#' * `acknowledgements` acknowledgement text, limited to 250 words
+#' * `bibliography` BibTeX `.bib` file
+#' * `longtable` set to `true` to include the `longtable` package, used by default from `pandoc` to convert markdown to LaTeX code
+#' * `header-includes`: custom additions to the header, before the `\begin{document}` statement
+#' * `include-after`: for including additional LaTeX code before the `\end{document}` statement
+#' @export
+#' @rdname article
+isba_article <- function(..., keep_tex = TRUE, highlight = NULL, citation_package = "natbib") {
+  if (citation_package != "natbib") {
+    stop("ISBA template only supports `natbib` for citation processing.")
+  }
+  # from https://github.com/rstudio/rmarkdown/issues/372
+  # md_extensions <- c("+ascii_identifiers", "+tex_math_single_backslash", "-autolink_bare_uris")
+  if (!rmarkdown::pandoc_available("2.10")) {
+    stop("`isba_article()` requires a minimum of pandoc 2.10.")
+  }
+  pdf_document_format(
+    "isba",
+    keep_tex = keep_tex, highlight = highlight, citation_package = citation_package, ...
   )
 }