+++ R/ieee_article.R
@@ -1,7 +1,7 @@
 #' IEEE Transactions journal format.
 #'
 #' Format for creating submissions to IEEE Transaction journals. Adapted from
-#' <http://www.ieee.org/publications_standards/publications/authors/author_templates.html>.
+#' <https://www.ieee.org/publications_standards/publications/authors/author_templates.html>
 #'
 #' Presently, only the `"conference"` paper mode offered by the
 #' `IEEEtran.cls` is supported.
@@ -36,23 +36,20 @@
 #' <http://mirrors.rit.edu/CTAN/macros/latex/contrib/IEEEtran/IEEEtran_HOWTO.pdf>
 #' @importFrom rmarkdown pandoc_variable_arg
 #' @export
-ieee_article <- function(
-  draftmode   = c("final", "draft", "draftcls", "draftclsnofoot"),
-  hyphenfixes      = "op-tical net-works semi-conduc-tor",
-  IEEEspecialpaper = "",
-  with_ifpdf       = FALSE,
-  with_cite        = FALSE,
-  with_amsmath     = FALSE,
-  with_algorithmic = FALSE,
-  with_subfig      = FALSE,
-  with_array       = FALSE,
-  with_dblfloatfix = FALSE,
-  keep_tex         = TRUE,
-  pandoc_args = NULL,
-  md_extensions    = c("-autolink_bare_uris"),
-  ...
-) {
-
+ieee_article <- function(draftmode = c("final", "draft", "draftcls", "draftclsnofoot"),
+                         hyphenfixes = "op-tical net-works semi-conduc-tor",
+                         IEEEspecialpaper = "",
+                         with_ifpdf = FALSE,
+                         with_cite = FALSE,
+                         with_amsmath = FALSE,
+                         with_algorithmic = FALSE,
+                         with_subfig = FALSE,
+                         with_array = FALSE,
+                         with_dblfloatfix = FALSE,
+                         keep_tex = TRUE,
+                         pandoc_args = NULL,
+                         md_extensions = c("-autolink_bare_uris"),
+                         ...) {
   args <- c()
 
   draftmode <- match.arg(draftmode)
@@ -65,23 +62,24 @@ ieee_article <- function(
     args <- c(args, "IEEEspecialpaper" = IEEEspecialpaper)
   }
 
-  plist <- c("with_ifpdf"       = with_ifpdf,
-             "with_cite"        = with_cite,
-             "with_amsmath"     = with_amsmath,
-             "with_algorithmic" = with_algorithmic,
-             "with_subfig"      = with_subfig,
-             "with_array"       = with_array,
-             "with_dblfloatfix" = with_dblfloatfix)
+  plist <- c(
+    "with_ifpdf" = with_ifpdf,
+    "with_cite" = with_cite,
+    "with_amsmath" = with_amsmath,
+    "with_algorithmic" = with_algorithmic,
+    "with_subfig" = with_subfig,
+    "with_array" = with_array,
+    "with_dblfloatfix" = with_dblfloatfix
+  )
 
   args <- c(args, plist[plist])
 
   # Convert to pandoc arguments
-  pandoc_arg_list <- mapply(rmarkdown::pandoc_variable_arg, names(args), args,
-                            USE.NAMES = FALSE, SIMPLIFY = FALSE)
-  pandoc_arg_list <- unlist(pandoc_arg_list)
+  pandoc_arg_list <- list_to_pandoc_variable_args(args)
 
   pdf_document_format(
-    "ieee", pandoc_args = c(pandoc_arg_list, pandoc_args),
+    "ieee",
+    pandoc_args = c(pandoc_arg_list, pandoc_args),
     keep_tex = keep_tex, md_extensions = md_extensions,
     ...
   )