+++ R/rjournal_article.R
@@ -1,5 +1,13 @@
 #' R Journal format.
 #'
+#' @description
+#' `r lifecycle::badge('deprecated')`
+#'
+#' This function is now deprecated in favor of the [**rjtools**](https://rjournal.github.io/rjtools/) package
+#' which is now officialy recommanded by R Journal <https://rjournal.github.io/submissions.html>. See below for document
+#'
+#' @details # About this format and the R Journal requirements
+#'
 #' Format for creating R Journal articles. Adapted from
 #' <https://journal.r-project.org/submissions.html>.
 #'
@@ -7,7 +15,7 @@
 #' Journal_ style and information on how to prepare your article for submission,
 #' see the [Instructions for Authors](https://journal.r-project.org/share/author-guide.pdf)
 #'
-#' ## About this format and the R Journal requirements
+
 #'
 #' `rticles::rjournal_article` will help you build the correct files requirements:
 #'
@@ -50,12 +58,14 @@
 #' @inheritParams rmarkdown::pdf_document
 #' @param ... Arguments to [rmarkdown::pdf_document()].
 #' @export
-rjournal_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') {
+rjournal_article <- function(..., keep_tex = TRUE, citation_package = "natbib") {
 
-  rmarkdown::pandoc_available('2.2', TRUE)
+  lifecycle::deprecate_warn("0.25", "rjournal_article()", "rjtools::rjournal_pdf_article()", details = "See official recommandation at https://rjournal.github.io/submissions.html")
+  rmarkdown::pandoc_available("2.2", TRUE)
 
   base <- pdf_document_format(
-    "rjournal", highlight = NULL, citation_package = citation_package,
+    "rjournal",
+    highlight = NULL, citation_package = citation_package,
     keep_tex = keep_tex, ...
   )
 
@@ -92,8 +102,9 @@ rjournal_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') 
     # underscores in the filename will be problematic in \input{filename};
     # pandoc will escape underscores but it should not, i.e., should be
     # \input{foo_bar} instead of \input{foo\_bar}
-    if (filename != (filename2 <- gsub('_', '-', filename))) {
-      file.rename(filename, filename2); filename <- filename2
+    if (filename != (filename2 <- gsub("_", "-", filename))) {
+      file.rename(filename, filename2)
+      filename <- filename2
     }
 
     # Copy purl-ed R file with the correct name
@@ -128,7 +139,7 @@ rjournal_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') 
     # check bibliography name
     bib_filename <- metadata$bibliography
     if (length(bib_filename) == 1 &&
-        xfun::sans_ext(bib_filename) != xfun::sans_ext(filename)) {
+      xfun::sans_ext(bib_filename) != xfun::sans_ext(filename)) {
       msg <- paste0(
         "Per R journal requirement, bibliography file and tex file should",
         " have the same name. Currently, you have a bib file ", bib_filename,
@@ -140,9 +151,10 @@ rjournal_article <- function(..., keep_tex = TRUE, citation_package = 'natbib') 
 
     # Create RJwrapper.tex per R Journal requirement
     m <- list(filename = xfun::sans_ext(filename))
-    h <- get_list_element(metadata, c('output', 'rticles::rjournal_article', 'includes', 'in_header'))
-    h <- c(h, if (length(preamble <- unlist(metadata[c('preamble', 'header-includes')]))) {
-      f <- tempfile(fileext = '.tex'); on.exit(unlink(f), add = TRUE)
+    h <- get_list_element(metadata, c("output", "rticles::rjournal_article", "includes", "in_header"))
+    h <- c(h, if (length(preamble <- unlist(metadata[c("preamble", "header-includes")]))) {
+      f <- tempfile(fileext = ".tex")
+      on.exit(unlink(f), add = TRUE)
       xfun::write_utf8(preamble, f)
       f
     })